Cambiar Id de modelos de BIGINT a INTEGER

+++ app/models/categoria.model.js
@@ -1,21 +1,13 @@
 module.exports = (sequelize, Sequelize) => {
-    const Categoria = sequelize.define("Category", {
-        id: {
-            type: Sequelize.BIGINT,
-            primaryKey: true,
-            autoIncrement: true,
-        },
-        nombre:{
-            type: Sequelize.STRING,
-        },
-    });
-    return Categoria;
+  const Categoria = sequelize.define("Category", {
+    id: {
+      type: Sequelize.INTEGER,
+      primaryKey: true,
+      autoIncrement: true,
+    },
+    nombre: {
+      type: Sequelize.STRING,
+    },
+  });
+  return Categoria;
 };
-
-
-
-
-
-
-
-
