+++ app/models/venta.model.js
@@ -1,27 +1,19 @@
 module.exports = (sequelize, Sequelize) => {
-    const Venta = sequelize.define("Venta", {
-        cliente: {
-            type: Sequelize.STRING
-        },
-        total: {
-            type: Sequelize.BIGINT
-        },
-        factura: {
-            type: Sequelize.STRING
-        },
-        id: {
-            type: Sequelize.BIGINT,
-            primaryKey: true,
-            autoIncrement: true
-        }
-    });
-    return Venta;
+  const Venta = sequelize.define("Venta", {
+    cliente: {
+      type: Sequelize.STRING,
+    },
+    total: {
+      type: Sequelize.BIGINT,
+    },
+    factura: {
+      type: Sequelize.STRING,
+    },
+    id: {
+      type: Sequelize.INTEGER,
+      primaryKey: true,
+      autoIncrement: true,
+    },
+  });
+  return Venta;
 };
-
-
-
-
-
-
-
-
